trying to fix tests and minor renamings
