+++ works/templates/works.html
@@ -8,7 +8,7 @@
 
     {% if is_admin %}
       <p class="alert alert-info">
-        <strong>Admin view:</strong> You can see all publications regardless of status. Status labels are shown next to each entry.
+        <strong>Admin view:</strong> You can see all works regardless of status. Status labels are shown next to each entry.
       </p>
     {% endif %}
 
@@ -137,7 +137,7 @@
           </div>
         </div>
       {% empty %}
-        <p class="alert alert-warning">No publications found.</p>
+        <p class="alert alert-warning">No works found.</p>
       {% endfor %}
     </div>
